+++ VisioAutomation_2010/VisioAutomation/ShapeSheetQuery/CommonQueries/HyperlinkCellsQuery.cs
@@ -1,6 +1,6 @@
 using IVisio = Microsoft.Office.Interop.Visio;
 
-namespace VisioAutomation.ShapeSheetQuery.Common
+namespace VisioAutomation.ShapeSheetQuery.CommonQueries
 {
     class HyperlinkCellsQuery : CellQuery
     {
