+++ VisioAutomation_2010/VisioAutomation/VisioAutomation.csproj
@@ -79,19 +79,19 @@
     <Compile Include="Fonts\FontHelper.cs" />
     <Compile Include="Layers\LayerHelper.cs" />
     <Compile Include="ShapeSheetQuery\CellColumnType.cs" />
-    <Compile Include="ShapeSheetQuery\Common\CharacterFormatCellsQuery.cs" />
-    <Compile Include="ShapeSheetQuery\Common\ConnectionPointCellsQuery.cs" />
-    <Compile Include="ShapeSheetQuery\Common\ControlCellsQuery.cs" />
-    <Compile Include="ShapeSheetQuery\Common\CustomPropertyCellsQuery.cs" />
-    <Compile Include="ShapeSheetQuery\Common\HyperlinkCellsQuery.cs" />
-    <Compile Include="ShapeSheetQuery\Common\LockCellsQuery.cs" />
-    <Compile Include="ShapeSheetQuery\Common\PageCellsQuery.cs" />
-    <Compile Include="ShapeSheetQuery\Common\ParagraphFormatCellsQuery.cs" />
-    <Compile Include="ShapeSheetQuery\Common\ShapeFormatCellsQuery.cs" />
-    <Compile Include="ShapeSheetQuery\Common\ShapeLayoutCellsQuery.cs" />
-    <Compile Include="ShapeSheetQuery\Common\TextBlockCellsQuery.cs" />
-    <Compile Include="ShapeSheetQuery\Common\UserDefinedCellsQuery.cs" />
-    <Compile Include="ShapeSheetQuery\Common\XFormCellsQuery.cs" />
+    <Compile Include="ShapeSheetQuery\CommonQueries\CharacterFormatCellsQuery.cs" />
+    <Compile Include="ShapeSheetQuery\CommonQueries\ConnectionPointCellsQuery.cs" />
+    <Compile Include="ShapeSheetQuery\CommonQueries\ControlCellsQuery.cs" />
+    <Compile Include="ShapeSheetQuery\CommonQueries\CustomPropertyCellsQuery.cs" />
+    <Compile Include="ShapeSheetQuery\CommonQueries\HyperlinkCellsQuery.cs" />
+    <Compile Include="ShapeSheetQuery\CommonQueries\LockCellsQuery.cs" />
+    <Compile Include="ShapeSheetQuery\CommonQueries\PageCellsQuery.cs" />
+    <Compile Include="ShapeSheetQuery\CommonQueries\ParagraphFormatCellsQuery.cs" />
+    <Compile Include="ShapeSheetQuery\CommonQueries\ShapeFormatCellsQuery.cs" />
+    <Compile Include="ShapeSheetQuery\CommonQueries\ShapeLayoutCellsQuery.cs" />
+    <Compile Include="ShapeSheetQuery\CommonQueries\TextBlockCellsQuery.cs" />
+    <Compile Include="ShapeSheetQuery\CommonQueries\UserDefinedCellsQuery.cs" />
+    <Compile Include="ShapeSheetQuery\CommonQueries\XFormCellsQuery.cs" />
     <Compile Include="ShapeSheet\ShapeSheetHelper.cs" />
     <Compile Include="ShapeSheet\ShapeSheetSurface.cs" />
     <Compile Include="Drawing\DrawingSurface.cs" />
